refactor notifs to implement likes

+++ pages/api/get-notifications.ts
@@ -12,35 +12,48 @@ export default async function getNotificationsHandler(req: NextApiRequest, res: 
 
     try {
         const thisUser: User = await getCurrUserRequest(session.user.email);
-        let notifications: Notification[] = await getNotifications(thisUser._id.toString());
-        if (notifications.length === 0) return res.status(200).json({notifications: [], users: [], updates: [], updateUsers: []});
+
+        let notifications = await notificationModel.aggregate([
+            {$match: {userId: thisUser._id}},
+            {
+                $lookup: {
+                    from: "users",
+                    foreignField: "_id",
+                    localField: "authorId",
+                    as: "authorArr",
+                },
+            },
+            {
+                $lookup: {
+                    from: "updates",
+                    let: {"updateId": "$updateId"},
+                    pipeline: [
+                        {$match: {$expr: {$eq: ["$_id", "$$updateId"]}}},
+                        {
+                            $lookup: {
+                                from: "users",
+                                foreignField: "_id",
+                                localField: "userId",
+                                as: "userArr",
+                            }
+                        }
+                    ],
+                    as: "updateArr",
+                },
+            },
+        ]);
+
 
         // delete read notifications older than 14 days
         const oldNotificationIds: string[] = notifications
             .filter(d => d.read && +new Date() - +new Date(d.createdAt) > (14 * 8.64e7))
             .map(d => d._id.toString());
-        await notificationModel.deleteMany({"_id": {$in: oldNotificationIds}});
+        if (oldNotificationIds.length) await notificationModel.deleteMany({"_id": {$in: oldNotificationIds}});
 
         notifications = notifications.filter(d => !oldNotificationIds.includes(d._id.toString()))
 
-        const uniqueAuthorIds: string[] = notifications.map(d => d.authorId).filter((d, i, a) => a.indexOf(d) === i);
-        const uniqueUpdateIds: string[] = notifications.map(d => d.updateId).filter((d, i, a) => a.indexOf(d) === i);
-        const users: User[] = await userModel.find({ "_id": {$in: uniqueAuthorIds}});
-        const updates: Update[] = await updateModel.find({ "_id": {$in: uniqueUpdateIds}});
-        const uniqueUpdateUserIds: string[] = updates.map(d => d.userId).filter((d, i, a) => a.indexOf(d) === i);
-        const updateUsers: User[] = await userModel.find({ "_id": {$in: uniqueUpdateUserIds}});
-        return res.status(200).json({notifications: notifications, users: users, updates: updates, updateUsers: updateUsers});
+        return res.status(200).json({notifications: notifications});
     } catch (e) {
         res.status(500).json({error: e});
     }
-}
-
-export async function getNotifications(userId: string) {
-    await mongoose.connect(process.env.MONGODB_URL, {
-        useNewUrlParser: true,
-        useUnifiedTopology: true,
-        useFindAndModify: false,
-    });
-
-    return notificationModel.find({ userId: userId });
 }
